+++ src/main/java/net/telepathicgrunt/bumblezone/entities/BzEntities.java
@@ -1,7 +1,6 @@
 package net.telepathicgrunt.bumblezone.entities;
 
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.minecraft.SharedConstants;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.entity.SpawnRestriction;