+++ CHANGELOG.md
@@ -8,7 +8,10 @@ Enter the dimension by throwing an enderpearl at Bee Nest and exit it by going a
        | The Bumblezone Fabric changelog |
 
    (V.2.3.1 Changes) (1.16.5 Minecraft)
-   
+
+   Blocks:
+- Fixed Honey Crystal block so that it rotates and mirrors properly when loaded from nbt files.
+
    Features:
 - Optimized waterfall feature to use honeycombs_that_features_can_carve block tag.